+++ modules/ffm-plus/src/main/java/club/doki7/ffm/ptr/CLongPtr.java
@@ -10,9 +10,9 @@ import org.jetbrains.annotations.Nullable;
 
 import java.lang.foreign.Arena;
 import java.lang.foreign.MemorySegment;
-import java.util.Iterator;
 import java.util.List;
 import java.util.NoSuchElementException;
+import java.util.PrimitiveIterator;
 
 /// Represents a pointer to 32-bit integer(s) in native memory.
 ///
@@ -81,7 +81,7 @@ public record CLongPtr(MemorySegment segment) implements IPointer, Iterable<Long
         return new CLongPtr(segment.asSlice(0, end * NativeLayout.C_LONG_SIZE));
     }
 
-    public @NotNull Iterator<Long> iterator() {
+    public @NotNull PrimitiveIterator.OfLong iterator() {
         return new Iter(segment);
     }
 
@@ -123,7 +123,7 @@ public record CLongPtr(MemorySegment segment) implements IPointer, Iterable<Long
     }
 
     /// An iterator over the integers.
-    private static final class Iter implements Iterator<Long> {
+    private static final class Iter implements PrimitiveIterator.OfLong {
         Iter(@NotNull MemorySegment segment) {
             this.segment = segment;
         }
@@ -134,7 +134,7 @@ public record CLongPtr(MemorySegment segment) implements IPointer, Iterable<Long
         }
 
         @Override
-        public Long next() {
+        public long nextLong() {
             if (!hasNext()) {
                 throw new NoSuchElementException("No more elements to read");
             }