+++ modules/ffm-plus/src/main/java/club/doki7/ffm/ptr/IntPtr.java
@@ -13,9 +13,9 @@ import java.lang.foreign.ValueLayout;
 import java.nio.Buffer;
 import java.nio.IntBuffer;
 import java.util.Collection;
-import java.util.Iterator;
 import java.util.List;
 import java.util.NoSuchElementException;
+import java.util.PrimitiveIterator;
 
 /// Represents a pointer to 32-bit integer(s) in native memory.
 ///
@@ -91,7 +91,7 @@ public record IntPtr(@NotNull MemorySegment segment) implements IPointer, Iterab
     }
 
     @Override
-    public @NotNull Iterator<Integer> iterator() {
+    public @NotNull PrimitiveIterator.OfInt iterator() {
         return new Iter(segment);
     }
 
@@ -228,7 +228,7 @@ public record IntPtr(@NotNull MemorySegment segment) implements IPointer, Iterab
     }
 
     /// An iterator over the integers.
-    private static final class Iter implements Iterator<Integer> {
+    private static final class Iter implements PrimitiveIterator.OfInt {
         Iter(@NotNull MemorySegment segment) {
             this.segment = segment;
         }
@@ -239,7 +239,7 @@ public record IntPtr(@NotNull MemorySegment segment) implements IPointer, Iterab
         }
 
         @Override
-        public Integer next() {
+        public int nextInt() {
             if (!hasNext()) {
                 throw new NoSuchElementException("No more integers to read");
             }