+++ modules/ffm-plus/src/main/java/club/doki7/ffm/ptr/LongPtr.java
@@ -13,10 +13,9 @@ import java.lang.foreign.ValueLayout;
 import java.nio.Buffer;
 import java.nio.LongBuffer;
 import java.util.Collection;
-import java.util.Iterator;
 import java.util.List;
 import java.util.NoSuchElementException;
-
+import java.util.PrimitiveIterator;
 /// Represents a pointer to 64-bit long integer(s) in native memory.
 ///
 /// The property {@link #segment()} should always be not-null
@@ -92,7 +91,7 @@ public record LongPtr(@NotNull MemorySegment segment) implements IPointer, Itera
     }
 
     @Override
-    public @NotNull Iterator<Long> iterator() {
+    public @NotNull PrimitiveIterator.OfLong iterator() {
         return new Iter(segment);
     }
 
@@ -147,7 +146,7 @@ public record LongPtr(@NotNull MemorySegment segment) implements IPointer, Itera
     /// @param buffer the {@link LongBuffer} to use as the backing storage
     /// @return a new {@link LongPtr} that uses {@code buffer} as its backing storage
     /// @throws IllegalArgumentException if {@code buffer} is not direct, or its backing storage is
-    /// not properly alignd
+    /// not properly aligned
     public static @NotNull LongPtr checked(@NotNull LongBuffer buffer) {
         if (!buffer.isDirect()) {
             throw new IllegalArgumentException("Buffer must be direct");
@@ -217,7 +216,7 @@ public record LongPtr(@NotNull MemorySegment segment) implements IPointer, Itera
     }
 
     /// An iterator over the long integers.
-    private static final class Iter implements Iterator<Long> {
+    private static final class Iter implements PrimitiveIterator.OfLong {
         Iter(@NotNull MemorySegment segment) {
             this.segment = segment;
         }
@@ -228,7 +227,7 @@ public record LongPtr(@NotNull MemorySegment segment) implements IPointer, Itera
         }
 
         @Override
-        public Long next() {
+        public long nextLong() {
             if (!hasNext()) {
                 throw new NoSuchElementException("No more long integers to read");
             }